# [Enhancement] [vectorized] Runtime Filter support equivalent slot of outer join 